Fixes and new features.
- Swapped "Fase0" with "Fase1".
Now, when you press "Start" you’ll immediately be prompted to load the subtitles for processing. This reduces user interaction (previously, you had to wait for the original "Fase0" to finish before loading the subs, meaning you had to wait for it to complete, now it’s more immediate).

- "Fase6": Now, you’ll be asked whether you want to merge "On top.ass/.srt," "Comments.ass" and "Signs.ass" into the final file. If you choose not to merge them, the separate files will also be moved to the desktop.

The "On Top" lines will be inserted into "Final.ass/.srt" in timestamp order.

The "Comments" lines will be placed at the very top.

The "Signs" lines will be placed at the bottom, after the dialogue.

- If "On Top.ass/.srt," "Comments," and "Signs" are empty, they will no longer be moved to the desktop. (Previously, they were moved even if empty. The user would have deleted them afterward anyway.)

- "Fase0": Fixed the duplicate "[Events]" issue that occurred when separating "On Top.ass," "Comments.ass" and "Signs.ass," which also appeared in the "Final.ass" file.