Pyscene speed and precision up
Fase3:
- Improved scene change detection and reduced false positives (removed ContentDetector which only caused false positives with almost no benefit).

- Implemented video parallelization in pyscenedetect, making it significantly faster.
In short, based on the loaded subtitle lines, it will detect which parts of the video to discard and which parts to feed to pyscenedetect (without cutting the video, it relies on the timing of the lines).
Thus, not only does it reduce scene change detection time by not processing the entire video, but the process is also sped up by parallelizing only the relevant video segments.

- Implemented logic to prevent memory leaks during the process.

- The progress bar may now "go crazy" when "Fase3" starts, as multiple video segments are processed simultaneously at high speed, resulting in multiple active progress bars at the same time (this is not an issue if you see this behavior, it's normal).

Fase4:
- Improved the logic of a function that caused issues in previous versions. Now, with the new "Fase3" and the new "Fase4," you'll have fewer lines to adjust in the final file (of course, it’s not perfect and can still make mistakes. Remember that there are configurable values in the GUI adjust them according to your timing preferences).