New GUI setting and improvements
Gui:
- Added the new setting "Max gap 'empty' between two lines to attach (ms)".
This setting allows you to configure the threshold for deciding how much "empty" space there must be to merge two subtitle lines. Merging the lines improves fluidity between them (this was already possible before, with a default value of 300, but it was not configurable via the GUI).
If you leave it at 200, the lines won’t stretch too much to attach to the next line, but there will be a natural "break" based on the fluidity of the speech. (This is obviously a personal preference, but you can now adjust the threshold to your liking).

Fase4:
- Modified "gap_threshold" to make it configurable via the GUI, as mentioned earlier.
- Tweaked some default thresholds.

- Removed the modified function from v1.6 and reinstated the one from v1.5, which caused fewer errors.
- The new modified function, based on the v1.5 version, is now "smarter". It should make far fewer errors than before. (Errors still depend on the crappiness of the loaded subtitle timing and false positives/undetected scenes from pyscenedetect).

Fase2:
- Tweaked some default thresholds.

Config_Fase2.json:
- Adjusted some default thresholds.

Config_Fase4.json:
- Added the new parameter "gap_threshold".