Improved speech peak detection
Fase2:
- Changed the Db to detect audio peaks in speech.
It's a small change, but it significantly alters how peaks are detected.
The improvement mainly applies to low and nuanced speech note.
Obviously, it’s not perfect, but it should be far better than before.
This means you will have to find your preferred values ​​again if you had customized them in the GUI.

Config_Fase2.json:
- Adjusted default values to align with the new peak detection.