3 new GUI settings, improvements and more...
GUI.py:
- Improved the handling of the "Config" buttons in the GUI.
- Whisper commands are now managed outside the GUI.
- Added 3 new configurable settings in the GUI.

(Config Fase4)
1: "Max range to detect scene change before the initial timestamp (ms)".
Allows you to set the maximum distance to detect a scene change and link it before the initial timestamp of the line.

2: "Max range to detect scene change after the initial timestamp (ms)".
Allows you to set the maximum distance to detect a scene change and link it after the initial timestamp of the line.

3: (Config Whisper)
Allows you to change the language and select your preferred Whisper model.

Whisper ReTimer:
-Updated "Fase0", "Fase3", and "Fase4" with the latest changes in "Auto Sub ReTimer".

- Much more...